refactor: make vision module more modular
