+++ flygym/mujoco/vision/__init__.py
@@ -1,1 +1,2 @@
-from .retina import Retina, visualize_visual_input
+from .retina import Retina
+from .visualize import visualize_visual_input
